Add public_timestamp to mainstream schema
Public timestamp exists already for whitehall indexes.  We want to be
able to have a comparable timestamp, holding the date of the most recent
major update to a page, for documents from publisher, too (so that we
can put them in the correct order in the 'latest' feed pages).  This
will be populated by future pull requests.