Always initialize 'doc' variable as a Hash
In the UnifiedSearchPresenter, the current behaviour assumed that
the "fields" key returned for an Elasticsearch document is always a 
hash.

However, if a request is made for only a small number of fields, and the
returned document contains none of them, then the 'fields' key is nil.
This causes an exception when we attempt to assign to the '_metadata' 
key, as 'doc' is nil.

This fixes this by using the double-pipe convention to fallback to an
empty hash, in the case that the value of 'fields' is nil.