Synonyms for all existing MAA regulatory articles
This is a fix for a common search pattern on MAA content: people search
for "RA1002", etc, but the documents are marked as "Regulatory article
(RA) 1002".  Without these synonmys, the appropriate documents therefore
don't match.

This is a poor solution longer term; we need to find a way to improve
the tokenisation so that these synonyms aren't needed.