Allow topics in search results
This removes a filter which hides documents of type "specialist_sector"
(which are now called topics) in search results.

This shouldn't be deployed before a corresponding change in frontend to
display a topics more nicely.

The unit tests have been updated to match the expected change, but I've
left the structure for checking that the correct filters are applied
because we're quite likely to add more filters back in future.