Allow facet example scope of "query"
Previously, when requesting examples for facet values, the only allowed
scope has been "global" - ie, return examples for the facet field
regardless of whether they match the query or not.

This adds the option of "query" scope - ie, return only examples which
match the query and filters.

The motivation of doing this now is to allow us to make services and
information pages only list examples which are produced by the
organisation the services and information page is for.

For example, on the HMRC services and information page, with global
scope we have a "Maritime Safety" section for which none of the 4
examples come from HMRC.  With this scope, it would contain only
examples which are tagged with HMRC.

(Note - we'd also need to modify whitehall to request "query" scope to
turn this option on on the services and information pages.)