Add mainstream_browse_pages field
Mainstream browse pages were modelled as three separate fields: section,
subsection and subsubsection.  This is unhelpful in many ways (can't
search directly for all content in a subsection, can't facet usefully,
can't represent tagging to multiple browse pages), so model them instead
as a single field containing the full path.

If the mainstream_browse_pages field isn't populated in supplied data,
build it from the "section", "subsection" and "subsubsection" fields.

In future, we'll get them in this form directly, at which point we'll
also be able to there may be multiple browse pages tagged to a piece of
content.