Make RUMMAGER_INDEX=all refer to all indexes
Currently, it only refers to all indexes holding content.  This is
frequently confusing people.  Instead, make RUMMAGER_INDEX=all refer to
all indexes used by rummager.

This means that auxiliary indexes are also migrated when a
deploy:migrations is done - currently, they're not being.  This means
that if we changed the schemas of the auxiliary indexes, they wouldn't
be updated automatically, which also seems undesirable.