Extract entities when indexing a document
When a document is being indexed, call out to the entity extractor service.

If any entities are found, these are stored in the 'entities' field of the
document in elastic search.

If no entities are found, nothing is stored.

We define the entity extractor client in the `SearchConfig` which allows it
to be swapped out if needed.