Avoid calling entity-extractor when migrating indexes
We estimated that calling out to the entity extractor would add about 20
minutes to an index migration, which is obviously not desirable.

Entity extraction will happen when documents are added individually to
the index.

I'll also add a rake task to allow an index migration including entity
extration which could be used to bulk-update the index.