Option to migrate_index to enable entity extraction
Entity extraction is relatively slow on a full index. Locally it took
8ms per document which sounds quick but with 100k documents that's 13
minutes.

Therefore we don't want to do that normally when migrating indexes,
because the entities won't have changed.

However we may occasionally want to do a full entity extraction on all
documents, so this option would provide a way to do that.