Refactor error swallowing implementation to use decoration
Making a connection error swallower be a decorator feels like a cleaner
implementation.

The choice of whether or not to have error swallowing behaviour is made by
composing objects rather than setting options or flags. This means that the
underlying EntityExtractorClient does not need any knowledge of how to swallow
errors, or even to be aware of the configuration option, which dramatically
simplifies the implementation and corresponding tests.

The tests of the error swallower class are also simpler and more focused.

The intention of the configuration setup is also clearer because each different
object type is given a meaningful name (`standard_entity_extractor`,
`error_swallowing_entity_extractor`, `null_entity_extractor`).