Use entities to boost search results
Call entity extractor from UnifiedSearchBuilder to identify the entities
in the user's query.  Add a boost filter for each entity found.

We've currently set a boost of 20 to make the change really obvious when
testing, but expect that a much lower value will be appropriate in
practice.  We'll experiment in various ways to find an appropriate
value.

Still to do:
 - [ ] add end-to-end integration test of entity extraction
 - [ ] add debug option to control whether entity extraction happens
 - [ ] fina appropriate value for boost