Set enable_entity_extraction option in initializer
The logic around whether entity extraction is enabled was somewhat
confusing (and probably wrong). Instead, always configure whether entity
extraction is enabled in `config/initializers/entity_extraction.rb`,
defaulting this to true in development and test environments and false
in production.

Deployment to production can then overwrite this with whatever value is
appropriate for the deployment target.