Support filtering to docs with no value in a field
This adds support for a "special" value of `_MISSING` which can be
specified in a filter to cause documents which have no value in the
filter field to be returned.  For example:

    filter_specialist_sectors=_MISSING

would filter the search results to include only those documents which
have no value in the `specialist_sectors` field.

It can be combined with searches for actual values. eg parameters of:

    filter_specialist_sectors[]=_MISSING&filter_specialist_sectors[]=farming

would return documents which either had no value, or had a value of
`farming`.

This is primarily of use when using the search index to inspect the
contents of the index, for example using the
[content-explorer](https://github.com/alphagov/govuk-content-explorer)
tool.

The code changes are fairly straightforward.  I've added tests for
parameter parsing, and an integration test to ensure that the generated
searches work correctly.

I've also renamed the field `topics` in some existing tests (and
corresponding fixture generation code) to `specialist_sectors`, because
`topics` isn't actually a field which filtering is allowed on in our
live schema, but specialist_sectors is.