Group facet example searches
When calculating examples for facet searches (eg, with queries such as
facet_organisations=100,example_scope:query,examples:10), we use
elasticsearch's `msearch` API to send a query for each facet value to
gather the examples. This is fine when there's a reasonably small number
of facet values to fetch examples for, but can cause problems if too
many facet values are found.

There are two problems which may occur:

 - Elasticsearch has an internal queue limit on the number of searches
   to be performed: this defaults to 1000.  If we go close to this
   limit, we risk getting error responses saying that the queue is full.
   I believe the 1000 limit is shared across all nodes, so each query
   submitted by a user counts several times against the 1000 limit. I
   heave experimentally verified that it is possible to hit the limit
   without actually sending as many as 1000 queries simultaneously.

 - Elasticsearch may take too long to perform the individual queries
   involved in the `msearch` call and time out.

For our current database, the most pathological field is the
organisations field, which can trigger nearly 900 searches for examples
in the worst case.  This takes 9 seconds to run on my dev VM, and times
out on the live site (when called by accessing the publicly exposed
search API directly).  There are no "user-facing" pages on the live site
which trigger such pathological cases, though, because facet examples
are only requested for highly filtered searches.

To avoid these problems, instead of sending all the searches at once, we
send them in batches of 50.  This means that we wait for 50 queries to
have been fully completed before sending the next 50, which avoids both
the queue size problem and the timeout problem.

In future, this workaround will be insufficient to return results fast
enough.  An alternative approach would be to use a custom elasticsearch
"aggregation" to gather information on the facet examples in a single
pass through the index.  This ought to be many times faster, but is
dependent on us upgrading the version of elasticsearch in use in
production (we're currently on a version which doesn't support
aggregations).