Allow the scope of facets to be controlled
Currently, when we compute the possible options to include in a facet
field, we ask elasticsearch to add up totals for all documents which
match both the query and "most" of the filters.  Specifically, we
exclude any filters which are set on the field that we're asking for
facet options for.

This is a good default when we're using faceting to populate a
multi-select facet box (such as in our finder interfaces).  However,
it's not a good default if we're trying to provide navigation options
within documents in a subset of the site.  This is exposed, for example,
in the "content explorer" app at
http://govuk-content-explorer.herokuapp.com/ which uses the facets in
the search API to provided details of the variety of options: in that
app, when a filter is applied to a field, it's unhelpful and confusing
that the facet options for that field are calculated without the filter
having been applied.

This PR adds an option to the facet parameters to allow the scope of the
requested facet to be set to either "exclude_field_filter" (the default,
and the current behaviour), or "all_filters" to cause all filters to be
applied when calculating facet options.

For example, if there's no search, but there is a filter on
`organisations` of `hm-revenue-customs`, when requesting a facet on the
`organisations` field the default behaviour would be to return all
possible values of `organisations` in the facet result.  With `scope`
set to `all_filters`, the behaviour would be to return only
organisations for which there is some content tagged to both that
organisation and to `hm-revenue-customs`.

This interface is designed such that other options could be added in
future if desired (a plausible one is "no_filters" to compute facets
only on the "query" part of the search).

For more on the concept of facet scopes, see
http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search-facets.html#_scope