Populate `format` field for all documents
For non-edition document types, we've not been setting a value for the
`format` field.  This makes various things awkward (analysis, switching
presenters based on a document type but falling back to switching on
format, etc.).

Instead, add `format` to the "core" fields (ie, defined for every
document type), and set the value of `format` for any non-edition
document type to be the same as the document type.  As we evolve the
index to avoid having a monolithic "edition" type we'll keep updating
this field, until when we finally remove the edition type in favour of
separate document types for each format, the format field will always
have the same value as the document type.  At that point, we'll remove
the `format` field in favour of using the document type.

Note on testing: I considered adding a specific test case for this to
the `test/unit/elasticsearch_index_test.rb`, but decided against it at
present.  This test file needed to be altered in several places to
expect the correct value in the format field, so the code is covered.
However, `test/unit/elasticsearch_index_test.rb` is a very crufty file
which is in heavy need of refactoring - adding a test to this file
wouldn't increase coverage, but would make it harder to refactor the
test file.