Reduce code duplication of all FooRegistry classes
There are a number of `FooRegistry` classes like `TopicRegistry` which
duplicate a lot of code. To make it easier for others to read and
understand all of these classes, move them all into a single module
`Registry`, where they can inherit from a common `BaseRegistry` class.

That way, the differences and similarities between all of these classes
can be made clearer.