Handle response to _aliases from elasticsearch 1.0+
The response from the _aliases endpoint has changed slightly in
elasticsearch 1.0+, such that it only contains an "aliases" key for
indexes which have at least one alias.

This commit updates the index_group code to understand the new response
format (while continuing to work with the old format where all indexes
had an "aliases" key in the response).