Reduce the number of JSON libraries.
Rummager was previously using a mixture of the json gem, yajl and
multi_json.  This PR replaces all of these with the json gem that's
bundled as part of the ruby standard library.  For ruby 2.1.4, this is
version 1.8.1.

This also removes multi_json, as it doesn't make much sense in an
application.