Add parsing of field definitions
This adds a `field_definitions.json` file to the schema configuration
directory, which defines the configuration for all the fields we
currently use in our indexes.  It also adds code and tests for parsing
these into a FieldDefinitions object, which contains a FieldDefinition
struct for each defined type.

The configuration for the fields is essentially identical to the
existing configuration, except for some pieces of invalid configuration
which have been omitted (eg, the old configuration for date fields often
had the property `"index": "not_analyzed"`, but this is not a valid
value for the `index` property for date fields (and is ignored by
elasticsearch).