Add parsing of document types
This adds files to the schema configuration directory which define which
fields are contained in each document type, and which define allowed
values for the fields in a specific document type.

It adds a `base_document_type.json` file to define which fields exist in
all document types.  Specific document types are defined by JSON files
in the `document_types` directory, where the name of the file defines
the name of the document type.  The configuration added in this commit
matches that currently defined by the existing fields in the
`default/doctypes/` schema directory.

It also adds code and tests for parsing
these into a DocumentType objects, using a DocumentTypesParser which
produces a hash from type name to the DocumentType object for that type.