Add parsing of index schema configuration
This adds files to the schema configuration directory which define which
document types are contained in each search index.  Each index is
defined by a file in the `config/schema/indexes` directory, named
according to the name of the index.

It also adds code and tests for parsing a file into an IndexSchema
object, and for producing a hash from index name to the corresponding
IndexSchema.