Add IndexForSearch
This contains the necessary methods for performing a search using the
"unified_search" index.

Before this change, the `Index` object is used for both searching and
indexing.  However, when searching with the "unified_search" endpoint,
multiple elasticsearch indexes are searched.  This was implemented in a
fairly hacky way previously, but now that we model schemas in a more
structured way it becomes very awkward to make most methods of the
`Index` object work with multiple elasticsearch indexes.

Instead, we separating the methods used by the `unified_search` endpoint
into a separate object which doesn't have to handle any index-time
stuff.  Only this object then needs to work with multiple underlying
elasticsearch indexes.

The object exposes the `index_names` and `schema` properties, to allow
classes using this object to look up the allowed values for a particular
document type in a particular index - this is used in expanding fields,
in particular to support the finders.