Fix integration + functional tests for new schema
Mostly, this involves stopping trying to generate stub index
configuration, and just using the live configuration.  This makes the
tests into tests for the code-paths we can actually hit anyway.

Also, change tests which used the "rummager_test" index to use
"mainstream_test", since no configuration is now defined for "rummager"
indexes.

A couple of tests are removed:
 - `test_create_index_with_custom_mappings` tested that we could add
   extra fields to an index, but we never do this.
 - `test_should_be_able_to_index_a_document_with_additional_fields`
   was similar.

`elasticsearch_advanced_search_test` is updated to use named examples of
a `boolean` and a `date` field, instead of creating special fields for
testing that these work.

`elasticsearch_migration_test` is the only test that actually still
stubs schemas, because it's checking that changes in the schema work as
expected.