Only override redis namespace in tests.
This config was added in 6abaf7f to avoid any overspill development and
testing.  The problem is that we're not using the environment suffix in
production at the moment, therefore if this was deployed as is, it would
lose any jobs currently in the queue when the prefix was changed.

This hasn't been a problem up to now because we're overwriting this
initializer at deploy time.  In future, we'll move to only overwriting
the redis.yml file instead.