Replace the FieldDefinitions object with a hash
The FieldDefinitions object was just a wrapper around a hash, with a
`.get` method to look up values, and a `.es_config` method which was
only used during parsing. This is fairly pointless code anyway, but when
making use of the definitions it would be easier if we just had a hash,
so this commit replaces the FieldDefinitions object with a hash.