Add metasearch and page-traffic to the schema
Previously, the `best_bet` and `page-traffic` document types, and the
`metasearch` and `page-traffic` indexes, were not modelled in the
new-style schemas.  They need to be to allow them to be indexed when we
start using the new schema stuff at indexing time.

This requires allowing document types to indicate in their schema
whether they should inherit from the base document type.  All our
content documents should inherit from this, but best bets are a special
case and shouldn't.  I've modelled this by adding a "use_base_type"
field in the `document_type` JSON, which can be set to `false` to
disable inheriting from the standard base document type.  If we wish to
extend this in future to a more flexible inheritance model, we could
support naming a type to inherit from in this field, but we don't need
such complexity at present.