Populate 'public_timestamp' for all documents
specialist-publisher sets 'last_update' as the timestamp signifying the
freshness of a document. This is semantically identical to 'public_timestamp',
which is the field that other publishing applications use. To ease the
transition to using 'public_timestamp' across all applications (and removing
'last_update'), 'public_timestamp' will be set on any documents that are indexed
with 'last_update'. A re-index will update all existing documents.