Add @document_types to Index
This will allow the allowed document types to be checked when documents
are added to the index.

Previously, Index objects were often constructed using their elasticsearch
index name: this is generally something with a unique ID and a
timestamp as well as the underlying name.  This makes it hard to look up
the schema for the index reliably, so I've added an additional
constructor parameter to Index which should be set to the "real" name.

(Slightly confusingly, there is also already a "real_name" method which
makes a call to elasticsearch to look up the alias name.  I'm not going
going to attempt to tidy that up at present.)