Pass schema information when making Documents
Index.documents_by_format requires the list of allowed fields to be
specified.  This change updates Registry.new to take an additional
parameter containing the list of all field definitions, which are then
filtered to the ones that are wanted, and updates app.rb to pass these
parameters.

Document.from_hash now requires the list of document types allowed in
the index.  Update Index and ResultSet to pass this data in.