Update schema to say attachments are multivalued
I'd accidentally modelled attachments as not being multivalued.

Nothing else was using the "object" type, so I've simply renamed it to
"objects" and made it multivalued.

While I was at it, I've removed the `"multivalued": false` lines for the
best_bet field types, for consistency with all the other field types
where we're not explicit about using the default.