Fix handling of opaque object fields
When returning documents, Document.get is called for each field in the
document.  This was applying Array() to the value of each field, to
force it into an array, to assist with compatibility with elasticsearch
1.0+ (which always return arrays).

This works fine for scalar values like strings and integers, but for a
hash it converts the hash into an array of "key-value" arrays.  eg,

    { "foo" => "bar" }

becomes

    [ ["foo", "bar"] ]

This breaks things that use the "metadata" field quite badly, so instead
we need to check for anything other than Array or nil, and simply wrap
such things in an array.

This code can be tidied up once we've migrated to elasticsearch 1.0+
fully.