Use multivalued attribute in result_set_presenter
This makes rummager return single values for non-multivalued fields.
This code path is used by the "unified_search" endpoint (which doesn't
use the Document class to format its output).

Special fields (like "_index" or "_type") are always returned as single
values, so we don't need to convert them.

This looks up information in the schema, and the simplest way to provide
that in the unified_searcher_test is to use the information from
"sample_config", so I've also removed some stubbing from
unified_searcher_test in favour of that approach.