Don't calculate dates in elasticsearch
Previously, we've been using the `time()` function inside a script in
elasticsearch queries to get the current time.  This is used to apply a
boost to newish documents.

There are many problems with this approach:

 - It breaks with elasticsearch 1.3 onwards, because mvel is deprecated
   as the default language for scripts (in favour of groovy), and
   `time()` doesn't exist in groovy.
 - Calling time() is relatively expensive, and it shouldn't vary
   significantly as the query progresses anyway.
 - Calling time in the query like this makes it impossible for
   elasticsearch to cache the computed scores.

Instead, we calculate the time in ruby, and pass it to elasticsearch as
a parameter.  This makes the query work with all versions of
elasticsearch (at least, from 0.90 onwards, which is all we care about).

We also only pass the time to an accuracy of 60 seconds, to allow
elasticsearch to cache the computed scores for a minute.  (We don't
actually turn this caching on, so I don't think elasticsearch will do
it, but the option is available for future performance tuning.)

This change introduces timecop, so that we don't have non-deterministic
tests.