Allow for BulkIndexFailure error in closing test
With elasticsearch 1.4.3+, when we try and do a bulk write to a closed
index, we get a "success" response code, but with error reports for the
individual documents which fail (rather than the 0.90.x behaviour of
returning an HTTP 4xx error).  Update the test of writing to a closed
index to expect this as one of the possible responses.

This should work with both elasticsearch 1.4.3+ or 0.90.  It fails with
1.4.2, because that version has a bug which causes an internal
NullPointerException, and returns an HTTP 500 status code.