Rename some field types
Renames the "searchable_date" field type to just "date", and the
"stored_float" field type to just "float".

There are no other types of date or float fields at present, so it's not
particularly helpful to try and be more specific about these names.  If
we need to add separate configuration for only some date fields in
future (probably for performance reasons), we should change the names to
indicate the ways in which the type can be used at that point, but
currently there's no need to add this level of detail.