Allow multiple identifiers for specialist fields
Specialist documents published through Specialist Publisher can often
have multiple values for one field. Rather than try to police
acceptable data in Rummager, let's leave that to the publishing tool and
be as open as possible by default.

It's unlikely that many of the `*_type` and `*_state` fields could
reasonably have multiple values, but let's lock that down in the future
if it become necessary.