Add support for "reject"-style filters
For example:

  /unified_search.json?reject_organisations=hm-revenue-customs

would return all documents, except for those tagged with the
`hm-revenue-customs` organisation.  This is wanted for the new search
interface for manuals, where the results filtered to the manual are
to be shown in one area, and all the other results in a separate area.

This is implemented by adding a "reject" parameter to the Filter object,
which tracks whether the filter is a rejection filter or not.

If a user supplies both "reject_organisations" and
"filter_organisations", both the filters need to match.  This will
usually not be intended behaviour, but it might be convenient to write a
frontend application this way, so I haven't made it into a validation
failure.