Parse synonyms to generate appropriate config
In order to fix problems with synonyms, we need to apply different
filters at search and indexing time, to map words and phrases in the
documents to special terms representing synonyms.

This commit adds a parser for the existing synonyms data, and a `Synonyms`
class which is produced by this parser, and knows how to generate the
appropriate elasticsearch configuration for the synonyms.