Use separate "subfields" for synonyms
Rather than apply synonyms to the main text field as part of the
analysis chain, this makes elasticsearch make a copy of any
"searchable_text" field, using the "fields" feature of elasticsearch
(available in 1.0+ only: see
http://www.elastic.co/guide/en/elasticsearch/reference/current/mapping-core-types.html#_multi_fields_3).
Synonyms are then applied to this copy at both index and search time, to
map the words in a synonym group to a special identifier.  We use
identifiers of the form "!S<number>", where number is an integer, to
avoid any risk of conflicting with normal terms (which cannot contain
punctuation, or uppercase letters).

After applying the synonym mapping, we "protect" the synonym identifier
terms using a "synonym_protwords" filter, which marks the synonym terms
as keywords.  This prevents subsequent filters being applied to them, so
we can for example apply the stemming filters to the words which aren't
synonyms.

At search time, we run the query against both the "normal" field, and
the ".synonyms" field, and combine the weights.  This means that an
exact match will normally get a higher weight than a synonym match, but
that all synonyms in a group will be applied.