Make attachment metadata searchable in _all field
Editors have been associating metadata with attachments which helps
users to find those attachments.  These are identifiers, so shouldn't be
analysed as free text, but should be considered if a user performs a
free text search.

I've implemented this by adding a "searchable_identifier" type, and
using that for these types.  There may be some other fields which would
benefit from this type, but since this is a strict improvement that
fixes a problem we've had complaints about, I'd like to get this merged
and deployed quickly, rather than do a full audit.

We'll be continually reviewing the schema definitions in future, anyway,
to see how we can improve our overall ranking schemes.