Rename existing synonym support as "old_synonyms"
We're going to rework the existing synonym support, but we want to be
able to run the replacement in parallel with the existing support for a
while before turning the existing support off.  I'm confident that the
new approach will work better, but it requires the synonym lists being
manually reworked first, so this commit renames all references to the
existing support as "old_synonyms".  This will make it obvious which
bits to remove when we have switched over to the new support.

There are several specific changes:

 - Copy the existing `synonyms.yml` file to `old_synonyms.yml`.  The new
   synonym support will use the `synonyms.yml` file, so this lets that
   file be modified without affecting the "old synonyms".

 - We're going to stop using the `query_default` analyzer by default, so
   the name was misleading.  Instead, call it `query_with_old_synonyms`.
   Note that, to avoid errors during deploy, we still define the
   `query_default` analyzer, but no longer use it; it can be removed
   only once all app servers are running the new code.

 - Rename the elasticsearch filter used by this analyzer from `synonym`
   to `old_synonym`.

Note that the indexes must be migrated when this is deployed, to ensure
that the new `query_with_old_synonyms` analyzer exists.